refactor(tui): split form.go below 1000 lines
Move each form field type into a sibling file in package tui.
Keep Form layout and submit in form.go. No behavior change.

Related: #652
